Made colours top level for easier use
