+++ engine/src/main/kotlin/world/gregs/voidps/engine/client/ui/chat/ChatColour.kt
@@ -1,0 +1,11 @@
+package world.gregs.voidps.engine.client.ui.chat
+
+abstract class ChatColour(int: Int) : Colour(int)
+
+object ChatBlue : ChatColour(0x000080)
+object ChatRed : ChatColour(0x800000)
+object TradePurple : ChatColour(0x800080)
+object DuelBrown : ChatColour(0x7e3200)
+object AssistPurple : ChatColour(0x8824e3)
+object DropGreen : ChatColour(0x005100)
+object WarningRed : ChatColour(0x480000)
