+++ game/src/main/kotlin/world/gregs/voidps/world/activity/combat/consume/food/Kebab.kts
@@ -1,5 +1,5 @@
 import world.gregs.voidps.engine.client.message
-import world.gregs.voidps.engine.client.ui.Colour
+import world.gregs.voidps.engine.client.ui.chat.Red
 import world.gregs.voidps.engine.entity.character.forceChat
 import world.gregs.voidps.engine.entity.character.player.Player
 import world.gregs.voidps.engine.entity.character.player.chat.ChatType
@@ -39,7 +39,7 @@ on<Consume>({ item.id == "super_kebab" }) { player: Player ->
         val skill = Skill.all.filterNot { it == Skill.Constitution }.random()
         player.levels.drain(skill, multiplier = 0.05)
         player.message("That tasted very dodgy. You feel very ill.", ChatType.Filter)
-        player.message(Colour.Red { "world.gregs.voidps.world.activity.combat.consume.Eating the kebab has done damage to some of your stats." })
+        player.message(Red { "world.gregs.voidps.world.activity.combat.consume.Eating the kebab has done damage to some of your stats." })
     }
     cancel()
 }
